examples/claude-python/requirements.txt
claiv-memory>=0.6.4
anthropic>=0.25.0
python-dotenv>=1.0.0
